PDDA-474 Moving getCourtSites() out of HearingTypeService into AbstractService
